+++ src/renderer/src/components/terminal-pane/TerminalPane.tsx
@@ -1,33 +1,23 @@
 import { useEffect, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
 import type { CSSProperties } from 'react'
-import { TOGGLE_TERMINAL_PANE_EXPAND_EVENT } from '@/constants/terminal'
 import { useAppStore } from '../../store'
 import {
   DEFAULT_TERMINAL_DIVIDER_DARK,
   normalizeColor,
   resolveEffectiveTerminalAppearance
 } from '@/lib/terminal-theme'
-import { PaneManager } from '@/lib/pane-manager/pane-manager'
+import type { PaneManager } from '@/lib/pane-manager/pane-manager'
 import TerminalSearch from '@/components/TerminalSearch'
 import type { PtyTransport } from './pty-transport'
-import {
-  EMPTY_LAYOUT,
-  buildFontFamily,
-  serializeTerminalLayout,
-  replayTerminalLayout
-} from './layout-serialization'
-import {
-  createExpandCollapseActions,
-  restoreExpandedLayoutFrom,
-  applyExpandedLayoutTo
-} from './expand-collapse'
+import { EMPTY_LAYOUT, serializeTerminalLayout } from './layout-serialization'
+import { createExpandCollapseActions } from './expand-collapse'
 import { useTerminalKeyboardShortcuts, useTerminalFontZoom } from './keyboard-handlers'
-import { applyTerminalAppearance } from './terminal-appearance'
-import { connectPanePty } from './pty-connection'
 import TerminalContextMenu from './TerminalContextMenu'
-
-const CLOSE_ALL_CONTEXT_MENUS_EVENT = 'orca-close-all-context-menus'
+import { useSystemPrefersDark } from './use-system-prefers-dark'
+import { useTerminalPaneGlobalEffects } from './use-terminal-pane-global-effects'
+import { useTerminalPaneLifecycle } from './use-terminal-pane-lifecycle'
+import { useTerminalPaneContextMenu } from './use-terminal-pane-context-menu'
 
 type TerminalPaneProps = {
   tabId: string
@@ -46,8 +36,6 @@ export default function TerminalPane({
 }: TerminalPaneProps): React.JSX.Element {
   const containerRef = useRef<HTMLDivElement>(null)
   const managerRef = useRef<PaneManager | null>(null)
-  const contextPaneIdRef = useRef<number | null>(null)
-  const wasActiveRef = useRef(false)
   const paneFontSizesRef = useRef<Map<number, number>>(new Map())
   const expandedPaneIdRef = useRef<number | null>(null)
   const expandedStyleSnapshotRef = useRef<Map<HTMLElement, { display: string; flex: string }>>(
@@ -57,31 +45,26 @@ export default function TerminalPane({
   const pendingWritesRef = useRef<Map<number, string>>(new Map())
   const isActiveRef = useRef(isActive)
   isActiveRef.current = isActive
-  const [terminalMenuOpen, setTerminalMenuOpen] = useState(false)
-  const [terminalMenuPoint, setTerminalMenuPoint] = useState({ x: 0, y: 0 })
-  const menuOpenedAtRef = useRef(0)
+
   const [expandedPaneId, setExpandedPaneId] = useState<number | null>(null)
   const [searchOpen, setSearchOpen] = useState(false)
-  const setTabPaneExpanded = useAppStore((s) => s.setTabPaneExpanded)
-  const setTabCanExpandPane = useAppStore((s) => s.setTabCanExpandPane)
-  const savedLayout = useAppStore((s) => s.terminalLayoutsByTabId[tabId] ?? EMPTY_LAYOUT)
-  const setTabLayout = useAppStore((s) => s.setTabLayout)
+
+  const setTabPaneExpanded = useAppStore((store) => store.setTabPaneExpanded)
+  const setTabCanExpandPane = useAppStore((store) => store.setTabCanExpandPane)
+  const savedLayout = useAppStore((store) => store.terminalLayoutsByTabId[tabId] ?? EMPTY_LAYOUT)
+  const setTabLayout = useAppStore((store) => store.setTabLayout)
   const initialLayoutRef = useRef(savedLayout)
-  const updateTabTitle = useAppStore((s) => s.updateTabTitle)
-  const updateTabPtyId = useAppStore((s) => s.updateTabPtyId)
-  const clearTabPtyId = useAppStore((s) => s.clearTabPtyId)
-  const markWorktreeUnreadFromBell = useAppStore((s) => s.markWorktreeUnreadFromBell)
-  const settings = useAppStore((s) => s.settings)
+  const updateTabTitle = useAppStore((store) => store.updateTabTitle)
+  const updateTabPtyId = useAppStore((store) => store.updateTabPtyId)
+  const clearTabPtyId = useAppStore((store) => store.clearTabPtyId)
+  const markWorktreeUnreadFromBell = useAppStore((store) => store.markWorktreeUnreadFromBell)
+  const settings = useAppStore((store) => store.settings)
   const settingsRef = useRef(settings)
   settingsRef.current = settings
   const onPtyExitRef = useRef(onPtyExit)
   onPtyExitRef.current = onPtyExit
 
-  const [systemPrefersDark, setSystemPrefersDark] = useState(() =>
-    typeof window !== 'undefined' && typeof window.matchMedia === 'function'
-      ? window.matchMedia('(prefers-color-scheme: dark)').matches
-      : true
-  )
+  const systemPrefersDark = useSystemPrefersDark()
 
   const persistLayoutSnapshot = (): void => {
     const manager = managerRef.current
@@ -110,291 +93,35 @@ export default function TerminalPane({
     persistLayoutSnapshot
   })
 
-  const syncCanExpandState = (): void => {
-    const paneCount = managerRef.current?.getPanes().length ?? 1
-    setTabCanExpandPane(tabId, paneCount > 1)
-  }
-
-  const doApplyAppearance = (manager: PaneManager): void => {
-    const s = settingsRef.current
-    if (!s) {
-      return
-    }
-    applyTerminalAppearance(
-      manager,
-      s,
-      systemPrefersDark,
-      paneFontSizesRef.current,
-      paneTransportsRef.current
-    )
-  }
-
-  useEffect(() => {
-    const closeMenu = (): void => {
-      if (Date.now() - menuOpenedAtRef.current < 100) {
-        return
-      }
-      setTerminalMenuOpen(false)
-    }
-    window.addEventListener(CLOSE_ALL_CONTEXT_MENUS_EVENT, closeMenu)
-    return () => window.removeEventListener(CLOSE_ALL_CONTEXT_MENUS_EVENT, closeMenu)
-  }, [])
-
-  useEffect(() => {
-    const media = window.matchMedia('(prefers-color-scheme: dark)')
-    const handleChange = (event: MediaQueryListEvent): void => setSystemPrefersDark(event.matches)
-    setSystemPrefersDark(media.matches)
-    media.addEventListener('change', handleChange)
-    return () => media.removeEventListener('change', handleChange)
-  }, [])
-
-  // Initialize PaneManager instance once
-  useEffect(() => {
-    const container = containerRef.current
-    if (!container) {
-      return
-    }
-    let resizeRaf: number | null = null
-
-    const queueResizeAll = (focusActive: boolean): void => {
-      if (resizeRaf !== null) {
-        cancelAnimationFrame(resizeRaf)
-      }
-      resizeRaf = requestAnimationFrame(() => {
-        resizeRaf = null
-        const m = managerRef.current
-        if (!m) {
-          return
-        }
-        const panes = m.getPanes()
-        for (const p of panes) {
-          try {
-            p.fitAddon.fit()
-          } catch {
-            /* ignore */
-          }
-        }
-        if (focusActive) {
-          const active = m.getActivePane() ?? panes[0]
-          active?.terminal.focus()
-        }
-      })
-    }
-
-    let shouldPersistLayout = false
-    const ptyDeps = {
-      tabId,
-      worktreeId,
-      cwd,
-      paneTransportsRef,
-      pendingWritesRef,
-      isActiveRef,
-      onPtyExitRef,
-      clearTabPtyId,
-      updateTabTitle,
-      updateTabPtyId,
-      markWorktreeUnreadFromBell
-    }
-
-    const manager = new PaneManager(container, {
-      onPaneCreated: (pane) => {
-        doApplyAppearance(manager)
-        connectPanePty(pane, manager, ptyDeps)
-        queueResizeAll(true)
-      },
-      onPaneClosed: (paneId) => {
-        const transport = paneTransportsRef.current.get(paneId)
-        if (transport) {
-          transport.destroy?.()
-          paneTransportsRef.current.delete(paneId)
-        }
-        paneFontSizesRef.current.delete(paneId)
-        pendingWritesRef.current.delete(paneId)
-      },
-      onActivePaneChange: () => {
-        if (shouldPersistLayout) {
-          persistLayoutSnapshot()
-        }
-      },
-      onLayoutChanged: () => {
-        syncExpandedLayout()
-        syncCanExpandState()
-        queueResizeAll(false)
-        if (shouldPersistLayout) {
-          persistLayoutSnapshot()
-        }
-      },
-      terminalOptions: () => {
-        const cs = settingsRef.current
-        return {
-          fontSize: cs?.terminalFontSize ?? 14,
-          fontFamily: buildFontFamily(cs?.terminalFontFamily ?? 'SF Mono'),
-          scrollback: Math.min(
-            50_000,
-            Math.max(1000, Math.round((cs?.terminalScrollbackBytes ?? 10_000_000) / 200))
-          ),
-          cursorStyle: cs?.terminalCursorStyle ?? 'bar',
-          cursorBlink: cs?.terminalCursorBlink ?? true
-        }
-      },
-      onLinkClick: (url) => {
-        window.api.shell.openExternal(url)
-      }
-    })
-
-    managerRef.current = manager
-    const restoredPaneByLeafId = replayTerminalLayout(manager, initialLayoutRef.current, isActive)
-    const restoredActivePaneId =
-      (initialLayoutRef.current.activeLeafId
-        ? restoredPaneByLeafId.get(initialLayoutRef.current.activeLeafId)
-        : null) ??
-      manager.getActivePane()?.id ??
-      manager.getPanes()[0]?.id ??
-      null
-    if (restoredActivePaneId !== null) {
-      manager.setActivePane(restoredActivePaneId, { focus: isActive })
-    }
-
-    const restoredExpandedPaneId = initialLayoutRef.current.expandedLeafId
-      ? (restoredPaneByLeafId.get(initialLayoutRef.current.expandedLeafId) ?? null)
-      : null
-    if (restoredExpandedPaneId !== null && manager.getPanes().length > 1) {
-      setExpandedPane(restoredExpandedPaneId)
-      applyExpandedLayoutTo(restoredExpandedPaneId, {
-        managerRef,
-        containerRef,
-        expandedStyleSnapshotRef
-      })
-    } else {
-      setExpandedPane(null)
-    }
-    shouldPersistLayout = true
-    syncCanExpandState()
-    doApplyAppearance(manager)
-    queueResizeAll(isActive)
-    persistLayoutSnapshot()
-
-    return () => {
-      if (resizeRaf !== null) {
-        cancelAnimationFrame(resizeRaf)
-      }
-      restoreExpandedLayoutFrom(expandedStyleSnapshotRef.current)
-      for (const transport of paneTransportsRef.current.values()) {
-        transport.destroy?.()
-      }
-      paneTransportsRef.current.clear()
-      pendingWritesRef.current.clear()
-      manager.destroy()
-      managerRef.current = null
-      setTabPaneExpanded(tabId, false)
-      setTabCanExpandPane(tabId, false)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [tabId, cwd])
-
-  useEffect(() => {
-    const manager = managerRef.current
-    if (!manager || !settings) {
-      return
-    }
-    doApplyAppearance(manager)
-    const fontFamily = buildFontFamily(settings.terminalFontFamily)
-    for (const pane of manager.getPanes()) {
-      pane.terminal.options.fontFamily = fontFamily
-      try {
-        pane.fitAddon.fit()
-      } catch {
-        /* ignore */
-      }
-    }
-  }, [settings, systemPrefersDark])
+  useTerminalPaneLifecycle({
+    tabId,
+    worktreeId,
+    cwd,
+    isActive,
+    systemPrefersDark,
+    settings,
+    settingsRef,
+    initialLayoutRef,
+    managerRef,
+    containerRef,
+    expandedStyleSnapshotRef,
+    paneFontSizesRef,
+    paneTransportsRef,
+    pendingWritesRef,
+    isActiveRef,
+    onPtyExitRef,
+    clearTabPtyId,
+    updateTabTitle,
+    updateTabPtyId,
+    markWorktreeUnreadFromBell,
+    setTabPaneExpanded,
+    setTabCanExpandPane,
+    setExpandedPane,
+    syncExpandedLayout,
+    persistLayoutSnapshot
+  })
 
   useTerminalFontZoom({ isActive, managerRef, paneFontSizesRef, settingsRef })
-
-  useEffect(() => {
-    const manager = managerRef.current
-    if (!manager) {
-      return
-    }
-    if (isActive) {
-      manager.resumeRendering()
-      for (const [paneId, buf] of pendingWritesRef.current.entries()) {
-        if (buf.length > 0) {
-          const pane = manager.getPanes().find((p) => p.id === paneId)
-          if (pane) {
-            pane.terminal.write(buf)
-          }
-          pendingWritesRef.current.set(paneId, '')
-        }
-      }
-      requestAnimationFrame(() => {
-        const panes = manager.getPanes()
-        for (const p of panes) {
-          try {
-            p.fitAddon.fit()
-          } catch {
-            /* ignore */
-          }
-        }
-        const active = manager.getActivePane() ?? panes[0]
-        if (active) {
-          active.terminal.focus()
-        }
-      })
-    } else if (wasActiveRef.current) {
-      manager.suspendRendering()
-    }
-    wasActiveRef.current = isActive
-  }, [isActive])
-
-  useEffect(() => {
-    const onToggleExpand = (event: Event): void => {
-      const detail = (event as CustomEvent<{ tabId?: string }>).detail
-      if (!detail?.tabId || detail.tabId !== tabId) {
-        return
-      }
-      const manager = managerRef.current
-      if (!manager) {
-        return
-      }
-      const panes = manager.getPanes()
-      if (panes.length < 2) {
-        return
-      }
-      const pane = manager.getActivePane() ?? panes[0]
-      if (!pane) {
-        return
-      }
-      toggleExpandPane(pane.id)
-    }
-    window.addEventListener(TOGGLE_TERMINAL_PANE_EXPAND_EVENT, onToggleExpand)
-    return () => window.removeEventListener(TOGGLE_TERMINAL_PANE_EXPAND_EVENT, onToggleExpand)
-  }, [tabId])
-
-  useEffect(() => {
-    if (!isActive) {
-      return
-    }
-    const container = containerRef.current
-    if (!container) {
-      return
-    }
-    const ro = new ResizeObserver(() => {
-      const manager = managerRef.current
-      if (!manager) {
-        return
-      }
-      for (const p of manager.getPanes()) {
-        try {
-          p.fitAddon.fit()
-        } catch {
-          /* ignore */
-        }
-      }
-    })
-    ro.observe(container)
-    return () => ro.disconnect()
-  }, [isActive])
 
   useTerminalKeyboardShortcuts({
     isActive,
@@ -409,32 +136,16 @@ export default function TerminalPane({
     setSearchOpen
   })
 
-  useEffect(() => {
-    if (!isActive) {
-      return
-    }
-    const shellEscape = (p: string): string => {
-      if (/^[a-zA-Z0-9_./@:-]+$/.test(p)) {
-        return p
-      }
-      return `'${p.replace(/'/g, "'\\''")}'`
-    }
-    return window.api.ui.onFileDrop(({ path: filePath }) => {
-      const manager = managerRef.current
-      if (!manager) {
-        return
-      }
-      const pane = manager.getActivePane() ?? manager.getPanes()[0]
-      if (!pane) {
-        return
-      }
-      const transport = paneTransportsRef.current.get(pane.id)
-      if (!transport) {
-        return
-      }
-      transport.sendInput(shellEscape(filePath))
-    })
-  }, [isActive])
+  useTerminalPaneGlobalEffects({
+    tabId,
+    isActive,
+    managerRef,
+    containerRef,
+    pendingWritesRef,
+    paneTransportsRef,
+    isActiveRef,
+    toggleExpandPane
+  })
 
   // Intercept paste events on the terminal to bypass Chromium's native
   // clipboard pipeline. Chromium holds NSPasteboard references during format
@@ -479,79 +190,16 @@ export default function TerminalPane({
     return () => container.removeEventListener('paste', onPaste, { capture: true })
   }, [isActive])
 
-  const resolveMenuPane = () => {
-    const manager = managerRef.current
-    if (!manager) {
-      return null
-    }
-    const panes = manager.getPanes()
-    if (contextPaneIdRef.current !== null) {
-      const clickedPane = panes.find((p) => p.id === contextPaneIdRef.current) ?? null
-      if (clickedPane) {
-        return clickedPane
-      }
-    }
-    return manager.getActivePane() ?? panes[0] ?? null
-  }
+  const contextMenu = useTerminalPaneContextMenu({
+    managerRef,
+    paneTransportsRef,
+    toggleExpandPane
+  })
 
-  const handleCopy = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    const selection = pane.terminal.getSelection()
-    if (selection) {
-      await navigator.clipboard.writeText(selection)
-    }
-  }
-
-  const handlePaste = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    const text = await window.api.ui.readClipboardText()
-    if (text) {
-      pane.terminal.paste(text)
-    }
-  }
-
-  const handleSplitRight = (): void => {
-    const p = resolveMenuPane()
-    if (p) {
-      managerRef.current?.splitPane(p.id, 'vertical')
-    }
-  }
-  const handleSplitDown = (): void => {
-    const p = resolveMenuPane()
-    if (p) {
-      managerRef.current?.splitPane(p.id, 'horizontal')
-    }
-  }
-  const handleClosePane = (): void => {
-    const p = resolveMenuPane()
-    if (p && (managerRef.current?.getPanes().length ?? 0) > 1) {
-      managerRef.current?.closePane(p.id)
-    }
-  }
-  const handleClearScreen = (): void => {
-    const p = resolveMenuPane()
-    if (p) {
-      p.terminal.clear()
-    }
-  }
-  const handleToggleExpand = (): void => {
-    const p = resolveMenuPane()
-    if (p) {
-      toggleExpandPane(p.id)
-    }
-  }
-
-  const paneCount = managerRef.current?.getPanes().length ?? 1
-  const menuPaneId = resolveMenuPane()?.id ?? null
   const effectiveAppearance = settings
     ? resolveEffectiveTerminalAppearance(settings, systemPrefersDark)
     : null
+
   const terminalContainerStyle: CSSProperties = {
     display: isActive ? 'flex' : 'none',
     ['--orca-terminal-divider-color' as string]:
@@ -561,6 +209,7 @@ export default function TerminalPane({
       DEFAULT_TERMINAL_DIVIDER_DARK
     )
   }
+
   const activePane = managerRef.current?.getActivePane()
 
   return (
@@ -569,27 +218,7 @@ export default function TerminalPane({
         ref={containerRef}
         className="absolute inset-0 min-h-0 min-w-0"
         style={terminalContainerStyle}
-        onContextMenuCapture={(event) => {
-          event.preventDefault()
-          menuOpenedAtRef.current = Date.now()
-          window.dispatchEvent(new Event(CLOSE_ALL_CONTEXT_MENUS_EVENT))
-          const manager = managerRef.current
-          if (!manager) {
-            contextPaneIdRef.current = null
-            return
-          }
-          const target = event.target
-          if (!(target instanceof Node)) {
-            contextPaneIdRef.current = null
-            return
-          }
-          const clickedPane =
-            manager.getPanes().find((pane) => pane.container.contains(target)) ?? null
-          contextPaneIdRef.current = clickedPane?.id ?? null
-          const bounds = event.currentTarget.getBoundingClientRect()
-          setTerminalMenuPoint({ x: event.clientX - bounds.left, y: event.clientY - bounds.top })
-          setTerminalMenuOpen(true)
-        }}
+        onContextMenuCapture={contextMenu.onContextMenuCapture}
       />
       {activePane?.container &&
         createPortal(
@@ -601,20 +230,22 @@ export default function TerminalPane({
           activePane.container
         )}
       <TerminalContextMenu
-        open={terminalMenuOpen}
-        onOpenChange={setTerminalMenuOpen}
-        menuPoint={terminalMenuPoint}
-        menuOpenedAtRef={menuOpenedAtRef}
-        canClosePane={paneCount > 1}
-        canExpandPane={paneCount > 1}
-        menuPaneIsExpanded={menuPaneId !== null && menuPaneId === expandedPaneId}
-        onCopy={() => void handleCopy()}
-        onPaste={() => void handlePaste()}
-        onSplitRight={handleSplitRight}
-        onSplitDown={handleSplitDown}
-        onClosePane={handleClosePane}
-        onClearScreen={handleClearScreen}
-        onToggleExpand={handleToggleExpand}
+        open={contextMenu.open}
+        onOpenChange={contextMenu.setOpen}
+        menuPoint={contextMenu.point}
+        menuOpenedAtRef={contextMenu.menuOpenedAtRef}
+        canClosePane={contextMenu.paneCount > 1}
+        canExpandPane={contextMenu.paneCount > 1}
+        menuPaneIsExpanded={
+          contextMenu.menuPaneId !== null && contextMenu.menuPaneId === expandedPaneId
+        }
+        onCopy={() => void contextMenu.onCopy()}
+        onPaste={() => void contextMenu.onPaste()}
+        onSplitRight={contextMenu.onSplitRight}
+        onSplitDown={contextMenu.onSplitDown}
+        onClosePane={contextMenu.onClosePane}
+        onClearScreen={contextMenu.onClearScreen}
+        onToggleExpand={contextMenu.onToggleExpand}
       />
     </>
   )
