+++ src/renderer/src/lib/terminal-themes/defaults.ts
@@ -1,0 +1,54 @@
+import type { TerminalThemeMap } from './types'
+
+export const DEFAULT_TERMINAL_THEMES: TerminalThemeMap = {
+  // Exact colors from Ghostty source: src/terminal/color.zig + src/config/Config.zig
+  'Ghostty Default Style Dark': {
+    background: '#282c34',
+    foreground: '#ffffff',
+    cursor: '#ffffff',
+    cursorAccent: '#282c34',
+    selectionBackground: '#3e4451',
+    selectionForeground: '#ffffff',
+    black: '#1d1f21',
+    red: '#cc6666',
+    green: '#b5bd68',
+    yellow: '#f0c674',
+    blue: '#81a2be',
+    magenta: '#b294bb',
+    cyan: '#8abeb7',
+    white: '#c5c8c6',
+    brightBlack: '#666666',
+    brightRed: '#d54e53',
+    brightGreen: '#b9ca4a',
+    brightYellow: '#e7c547',
+    brightBlue: '#7aa6da',
+    brightMagenta: '#c397d8',
+    brightCyan: '#70c0b1',
+    brightWhite: '#eaeaea'
+  },
+
+  'Builtin Tango Light': {
+    background: '#ffffff',
+    foreground: '#2e3434',
+    cursor: '#2e3434',
+    cursorAccent: '#ffffff',
+    selectionBackground: '#accef7',
+    selectionForeground: '#2e3434',
+    black: '#2e3436',
+    red: '#cc0000',
+    green: '#4e9a06',
+    yellow: '#c4a000',
+    blue: '#3465a4',
+    magenta: '#75507b',
+    cyan: '#06989a',
+    white: '#d3d7cf',
+    brightBlack: '#555753',
+    brightRed: '#ef2929',
+    brightGreen: '#8ae234',
+    brightYellow: '#fce94f',
+    brightBlue: '#729fcf',
+    brightMagenta: '#ad7fa8',
+    brightCyan: '#34e2e2',
+    brightWhite: '#eeeeec'
+  }
+}
