+++ src/renderer/src/lib/terminal-themes/popular-dark-extended.ts
@@ -1,0 +1,220 @@
+import type { TerminalThemeMap } from './types'
+
+export const POPULAR_DARK_EXTENDED_TERMINAL_THEMES: TerminalThemeMap = {
+  'Material Dark': {
+    background: '#263238',
+    foreground: '#eeffff',
+    cursor: '#ffcc00',
+    cursorAccent: '#263238',
+    selectionBackground: '#546e7a',
+    selectionForeground: '#eeffff',
+    black: '#000000',
+    red: '#f07178',
+    green: '#c3e88d',
+    yellow: '#ffcb6b',
+    blue: '#82aaff',
+    magenta: '#c792ea',
+    cyan: '#89ddff',
+    white: '#eeffff',
+    brightBlack: '#546e7a',
+    brightRed: '#f07178',
+    brightGreen: '#c3e88d',
+    brightYellow: '#ffcb6b',
+    brightBlue: '#82aaff',
+    brightMagenta: '#c792ea',
+    brightCyan: '#89ddff',
+    brightWhite: '#ffffff'
+  },
+  'Ayu Dark': {
+    background: '#0a0e14',
+    foreground: '#b3b1ad',
+    cursor: '#e6b450',
+    cursorAccent: '#0a0e14',
+    selectionBackground: '#273747',
+    selectionForeground: '#b3b1ad',
+    black: '#01060e',
+    red: '#ea6c73',
+    green: '#91b362',
+    yellow: '#f9af4f',
+    blue: '#53bdfa',
+    magenta: '#fae994',
+    cyan: '#90e1c6',
+    white: '#c7c7c7',
+    brightBlack: '#686868',
+    brightRed: '#f07178',
+    brightGreen: '#c2d94c',
+    brightYellow: '#ffb454',
+    brightBlue: '#59c2ff',
+    brightMagenta: '#ffee99',
+    brightCyan: '#95e6cb',
+    brightWhite: '#ffffff'
+  },
+  Nightfox: {
+    background: '#192330',
+    foreground: '#cdcecf',
+    cursor: '#cdcecf',
+    cursorAccent: '#192330',
+    selectionBackground: '#2b3b51',
+    selectionForeground: '#cdcecf',
+    black: '#393b44',
+    red: '#c94f6d',
+    green: '#81b29a',
+    yellow: '#dbc074',
+    blue: '#719cd6',
+    magenta: '#9d79d6',
+    cyan: '#63cdcf',
+    white: '#dfdfe0',
+    brightBlack: '#575860',
+    brightRed: '#d16983',
+    brightGreen: '#8ebaa4',
+    brightYellow: '#e0c989',
+    brightBlue: '#86abdc',
+    brightMagenta: '#baa1e2',
+    brightCyan: '#7ad5d6',
+    brightWhite: '#e4e4e5'
+  },
+  Kanagawa: {
+    background: '#1f1f28',
+    foreground: '#dcd7ba',
+    cursor: '#c8c093',
+    cursorAccent: '#1f1f28',
+    selectionBackground: '#2d4f67',
+    selectionForeground: '#dcd7ba',
+    black: '#090618',
+    red: '#c34043',
+    green: '#76946a',
+    yellow: '#c0a36e',
+    blue: '#7e9cd8',
+    magenta: '#957fb8',
+    cyan: '#6a9589',
+    white: '#c8c093',
+    brightBlack: '#727169',
+    brightRed: '#e82424',
+    brightGreen: '#98bb6c',
+    brightYellow: '#e6c384',
+    brightBlue: '#7fb4ca',
+    brightMagenta: '#938aa9',
+    brightCyan: '#7aa89f',
+    brightWhite: '#dcd7ba'
+  },
+  'Rose Pine': {
+    background: '#191724',
+    foreground: '#e0def4',
+    cursor: '#524f67',
+    cursorAccent: '#e0def4',
+    selectionBackground: '#403d52',
+    selectionForeground: '#e0def4',
+    black: '#26233a',
+    red: '#eb6f92',
+    green: '#31748f',
+    yellow: '#f6c177',
+    blue: '#9ccfd8',
+    magenta: '#c4a7e7',
+    cyan: '#ebbcba',
+    white: '#e0def4',
+    brightBlack: '#6e6a86',
+    brightRed: '#eb6f92',
+    brightGreen: '#31748f',
+    brightYellow: '#f6c177',
+    brightBlue: '#9ccfd8',
+    brightMagenta: '#c4a7e7',
+    brightCyan: '#ebbcba',
+    brightWhite: '#e0def4'
+  },
+  'Everforest Dark': {
+    background: '#2d353b',
+    foreground: '#d3c6aa',
+    cursor: '#d3c6aa',
+    cursorAccent: '#2d353b',
+    selectionBackground: '#543a48',
+    selectionForeground: '#d3c6aa',
+    black: '#475258',
+    red: '#e67e80',
+    green: '#a7c080',
+    yellow: '#dbbc7f',
+    blue: '#7fbbb3',
+    magenta: '#d699b6',
+    cyan: '#83c092',
+    white: '#d3c6aa',
+    brightBlack: '#697379',
+    brightRed: '#e67e80',
+    brightGreen: '#a7c080',
+    brightYellow: '#dbbc7f',
+    brightBlue: '#7fbbb3',
+    brightMagenta: '#d699b6',
+    brightCyan: '#83c092',
+    brightWhite: '#d3c6aa'
+  },
+  Palenight: {
+    background: '#292d3e',
+    foreground: '#a6accd',
+    cursor: '#ffcc00',
+    cursorAccent: '#292d3e',
+    selectionBackground: '#434758',
+    selectionForeground: '#a6accd',
+    black: '#292d3e',
+    red: '#f07178',
+    green: '#c3e88d',
+    yellow: '#ffcb6b',
+    blue: '#82aaff',
+    magenta: '#c792ea',
+    cyan: '#89ddff',
+    white: '#a6accd',
+    brightBlack: '#676e95',
+    brightRed: '#f07178',
+    brightGreen: '#c3e88d',
+    brightYellow: '#ffcb6b',
+    brightBlue: '#82aaff',
+    brightMagenta: '#c792ea',
+    brightCyan: '#89ddff',
+    brightWhite: '#ffffff'
+  },
+  'Horizon Dark': {
+    background: '#1c1e26',
+    foreground: '#e0e0e0',
+    cursor: '#e95678',
+    cursorAccent: '#1c1e26',
+    selectionBackground: '#2e303e',
+    selectionForeground: '#e0e0e0',
+    black: '#16161c',
+    red: '#e95678',
+    green: '#29d398',
+    yellow: '#fab795',
+    blue: '#26bbd9',
+    magenta: '#ee64ac',
+    cyan: '#59e1e3',
+    white: '#d5d8da',
+    brightBlack: '#6c6f93',
+    brightRed: '#ec6a88',
+    brightGreen: '#3fdaa4',
+    brightYellow: '#fbc3a7',
+    brightBlue: '#3fc4de',
+    brightMagenta: '#f075b5',
+    brightCyan: '#6be4e6',
+    brightWhite: '#f0f0f0'
+  },
+  'Night Owl': {
+    background: '#011627',
+    foreground: '#d6deeb',
+    cursor: '#80a4c2',
+    cursorAccent: '#011627',
+    selectionBackground: '#1d3b53',
+    selectionForeground: '#d6deeb',
+    black: '#011627',
+    red: '#ef5350',
+    green: '#22da6e',
+    yellow: '#addb67',
+    blue: '#82aaff',
+    magenta: '#c792ea',
+    cyan: '#21c7a8',
+    white: '#ffffff',
+    brightBlack: '#575656',
+    brightRed: '#ef5350',
+    brightGreen: '#22da6e',
+    brightYellow: '#ffeb95',
+    brightBlue: '#82aaff',
+    brightMagenta: '#c792ea',
+    brightCyan: '#7fdbca',
+    brightWhite: '#ffffff'
+  }
+}
