+++ src/renderer/src/lib/terminal-themes/popular-light.ts
@@ -1,0 +1,203 @@
+import type { TerminalThemeMap } from './types'
+
+export const POPULAR_LIGHT_TERMINAL_THEMES: TerminalThemeMap = {
+  'Solarized Light': {
+    background: '#fdf6e3',
+    foreground: '#657b83',
+    cursor: '#657b83',
+    cursorAccent: '#fdf6e3',
+    selectionBackground: '#eee8d5',
+    selectionForeground: '#586e75',
+    black: '#073642',
+    red: '#dc322f',
+    green: '#859900',
+    yellow: '#b58900',
+    blue: '#268bd2',
+    magenta: '#d33682',
+    cyan: '#2aa198',
+    white: '#eee8d5',
+    brightBlack: '#002b36',
+    brightRed: '#cb4b16',
+    brightGreen: '#586e75',
+    brightYellow: '#657b83',
+    brightBlue: '#839496',
+    brightMagenta: '#6c71c4',
+    brightCyan: '#93a1a1',
+    brightWhite: '#fdf6e3'
+  },
+
+  'One Light': {
+    background: '#fafafa',
+    foreground: '#383a42',
+    cursor: '#526fff',
+    cursorAccent: '#fafafa',
+    selectionBackground: '#e5e5e6',
+    selectionForeground: '#383a42',
+    black: '#383a42',
+    red: '#e45649',
+    green: '#50a14f',
+    yellow: '#c18401',
+    blue: '#4078f2',
+    magenta: '#a626a4',
+    cyan: '#0184bc',
+    white: '#a0a1a7',
+    brightBlack: '#696c77',
+    brightRed: '#e45649',
+    brightGreen: '#50a14f',
+    brightYellow: '#c18401',
+    brightBlue: '#4078f2',
+    brightMagenta: '#a626a4',
+    brightCyan: '#0184bc',
+    brightWhite: '#fafafa'
+  },
+
+  'Catppuccin Latte': {
+    background: '#eff1f5',
+    foreground: '#4c4f69',
+    cursor: '#dc8a78',
+    cursorAccent: '#eff1f5',
+    selectionBackground: '#acb0be',
+    selectionForeground: '#4c4f69',
+    black: '#5c5f77',
+    red: '#d20f39',
+    green: '#40a02b',
+    yellow: '#df8e1d',
+    blue: '#1e66f5',
+    magenta: '#ea76cb',
+    cyan: '#179299',
+    white: '#acb0be',
+    brightBlack: '#6c6f85',
+    brightRed: '#d20f39',
+    brightGreen: '#40a02b',
+    brightYellow: '#df8e1d',
+    brightBlue: '#1e66f5',
+    brightMagenta: '#ea76cb',
+    brightCyan: '#179299',
+    brightWhite: '#bcc0cc'
+  },
+
+  'GitHub Light': {
+    background: '#ffffff',
+    foreground: '#24292e',
+    cursor: '#044289',
+    cursorAccent: '#ffffff',
+    selectionBackground: '#c8c8fa',
+    selectionForeground: '#24292e',
+    black: '#24292e',
+    red: '#d73a49',
+    green: '#28a745',
+    yellow: '#dbab09',
+    blue: '#0366d6',
+    magenta: '#5a32a3',
+    cyan: '#0598bc',
+    white: '#6a737d',
+    brightBlack: '#959da5',
+    brightRed: '#cb2431',
+    brightGreen: '#22863a',
+    brightYellow: '#b08800',
+    brightBlue: '#005cc5',
+    brightMagenta: '#5a32a3',
+    brightCyan: '#3192aa',
+    brightWhite: '#d1d5da'
+  },
+
+  'Rose Pine Dawn': {
+    background: '#faf4ed',
+    foreground: '#575279',
+    cursor: '#9893a5',
+    cursorAccent: '#faf4ed',
+    selectionBackground: '#dfdad9',
+    selectionForeground: '#575279',
+    black: '#f2e9e1',
+    red: '#b4637a',
+    green: '#286983',
+    yellow: '#ea9d34',
+    blue: '#56949f',
+    magenta: '#907aa9',
+    cyan: '#d7827e',
+    white: '#575279',
+    brightBlack: '#9893a5',
+    brightRed: '#b4637a',
+    brightGreen: '#286983',
+    brightYellow: '#ea9d34',
+    brightBlue: '#56949f',
+    brightMagenta: '#907aa9',
+    brightCyan: '#d7827e',
+    brightWhite: '#575279'
+  },
+
+  'Gruvbox Light': {
+    background: '#fbf1c7',
+    foreground: '#3c3836',
+    cursor: '#3c3836',
+    cursorAccent: '#fbf1c7',
+    selectionBackground: '#ebdbb2',
+    selectionForeground: '#3c3836',
+    black: '#fbf1c7',
+    red: '#cc241d',
+    green: '#98971a',
+    yellow: '#d79921',
+    blue: '#458588',
+    magenta: '#b16286',
+    cyan: '#689d6a',
+    white: '#7c6f64',
+    brightBlack: '#928374',
+    brightRed: '#9d0006',
+    brightGreen: '#79740e',
+    brightYellow: '#b57614',
+    brightBlue: '#076678',
+    brightMagenta: '#8f3f71',
+    brightCyan: '#427b58',
+    brightWhite: '#3c3836'
+  },
+
+  'Tokyo Night Light': {
+    background: '#d5d6db',
+    foreground: '#343b58',
+    cursor: '#343b58',
+    cursorAccent: '#d5d6db',
+    selectionBackground: '#9699a3',
+    selectionForeground: '#343b58',
+    black: '#0f0f14',
+    red: '#8c4351',
+    green: '#485e30',
+    yellow: '#8f5e15',
+    blue: '#34548a',
+    magenta: '#5a4a78',
+    cyan: '#0f4b6e',
+    white: '#343b58',
+    brightBlack: '#9699a3',
+    brightRed: '#8c4351',
+    brightGreen: '#485e30',
+    brightYellow: '#8f5e15',
+    brightBlue: '#34548a',
+    brightMagenta: '#5a4a78',
+    brightCyan: '#0f4b6e',
+    brightWhite: '#343b58'
+  },
+
+  'Everforest Light': {
+    background: '#fdf6e3',
+    foreground: '#5c6a72',
+    cursor: '#5c6a72',
+    cursorAccent: '#fdf6e3',
+    selectionBackground: '#e6e2cc',
+    selectionForeground: '#5c6a72',
+    black: '#5c6a72',
+    red: '#f85552',
+    green: '#8da101',
+    yellow: '#dfa000',
+    blue: '#3a94c5',
+    magenta: '#df69ba',
+    cyan: '#35a77c',
+    white: '#dfd9a8',
+    brightBlack: '#939f91',
+    brightRed: '#f85552',
+    brightGreen: '#8da101',
+    brightYellow: '#dfa000',
+    brightBlue: '#3a94c5',
+    brightMagenta: '#df69ba',
+    brightCyan: '#35a77c',
+    brightWhite: '#fdf6e3'
+  }
+}
