+++ src/renderer/src/lib/terminal-themes/types.ts
@@ -1,0 +1,3 @@
+import type { ITheme } from '@xterm/xterm'
+
+export type TerminalThemeMap = Record<string, ITheme>
